requirements.txt
streamlit
google-generativeai
edge-tts
ffmpeg-python
Pillow